Rename trellis-namespaces to trellis-namespace
