+++ components/namespace/bnd.bnd
