+++ components/namespace/src/main/java/module-info.java
@@ -13,11 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/**
- * Trellis Namespace Service using an in-memory map of namespaces
- *
- * <p>This package provides a mechanism for associating prefixes to namespaces.
- *
- * @author acoburn
- */
-package org.trellisldp.namespaces;
+module org.trellisldp.namespace {
+    exports org.trellisldp.namespace;
+
+    requires transitive org.trellisldp.api;
+    requires transitive org.trellisldp.vocabulary;
+
+    requires org.apache.commons.rdf.api;
+    requires jakarta.enterprise.cdi.api;
+    requires jakarta.inject;
+    requires microprofile.config.api;
+    requires org.slf4j;
+
+    provides org.trellisldp.api.NamespaceService
+        with org.trellisldp.namespace.SimpleNamespaceService;
+}
