+++ components/namespace/src/main/resources/META-INF/beans.xml
