+++ components/namespace/src/main/resources/META-INF/services/org.trellisldp.api.NamespaceService
@@ -1,0 +1,1 @@
+org.trellisldp.namespace.SimpleNamespaceService
