--- components/namespaces/src/main/resources/META-INF/services/org.trellisldp.api.NamespaceService
@@ -1,1 +1,0 @@
-org.trellisldp.namespaces.SimpleNamespaceService
