+++ platform/karaf/src/main/resources/features.xml
@@ -135,7 +135,7 @@
     <bundle>mvn:org.trellisldp/trellis-rdfa/${project.version}</bundle>
   </feature>
 
-  <feature name="trellis-namespaces" version="${project.version}">
+  <feature name="trellis-namespace" version="${project.version}">
     <details>Installs the Trellis JSON-file-based Namespace service</details>
 
     <feature dependency="true">trellis-api</feature>
@@ -144,7 +144,7 @@
     <bundle dependency="true">mvn:org.apache.servicemix.bundles/org.apache.servicemix.bundles.javax-inject/1_2</bundle>
     <bundle dependency="true">mvn:org.eclipse.microprofile.config/microprofile-config-api/${microprofileConfigVersion}</bundle>
 
-    <bundle>mvn:org.trellisldp/trellis-namespaces/${project.version}</bundle>
+    <bundle>mvn:org.trellisldp/trellis-namespace/${project.version}</bundle>
   </feature>
 
   <feature name="trellis-jdbc" version="${project.version}">
