+++ platform/quarkus/src/main/resources/application.properties
@@ -42,7 +42,7 @@ trellis.quarkus.authz-cache-size=1000
 trellis.quarkus.authz-cache-expire-seconds=600
 
 # Trellis namespaces
-trellis.namespaces.mapping=
+trellis.namespace.mapping=
 
 # Trellis file
 trellis.file.binary-path=data/binaries
