+++ settings.gradle
@@ -42,7 +42,7 @@ include ':trellis-dropwizard'
 include ':trellis-file'
 include ':trellis-io-jena'
 include ':trellis-jdbc'
-include ':trellis-namespaces'
+include ':trellis-namespace'
 include ':trellis-rdfa'
 include ':trellis-triplestore'
 include ':trellis-webdav'
@@ -80,7 +80,7 @@ project(':trellis-dropwizard').projectDir = new File(rootDir, "components/dropwi
 project(':trellis-file').projectDir = new File(rootDir, "components/file")
 project(':trellis-io-jena').projectDir = new File(rootDir, "components/io-jena")
 project(':trellis-jdbc').projectDir = new File(rootDir, "components/jdbc")
-project(':trellis-namespaces').projectDir = new File(rootDir, "components/namespaces")
+project(':trellis-namespace').projectDir = new File(rootDir, "components/namespace")
 project(':trellis-rdfa').projectDir = new File(rootDir, "components/rdfa")
 project(':trellis-test').projectDir = new File(rootDir, "components/test")
 project(':trellis-triplestore').projectDir = new File(rootDir, "components/triplestore")
